# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于 Gin + Gorm + Wire 的企业级后端服务",
    "title": "Sponge API",
    "termsOfService": "https://github.com/Mojitocean/sponge",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:18888",
  "basePath": "/gateway",
  "paths": {
...
    "/gateway/user/v1/login": {
      "post": {
        "description": "通过用户名密码登录并返回JWT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LoginReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/user/v1/profile": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取个人信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "获取个人信息",
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.UserInfoRes"
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "未授权或Token失效",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/user/v1/register": {
      "post": {
        "description": "根据用户名和密码创建新用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "用户注册",
        "parameters": [
          {
            "description": "注册信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.RegisterReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注册成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          },
          "400": {
            "description": "参数错误",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          },
          "500": {
            "description": "系统异常",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/user/v1/update/password": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "修改密码",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "修改密码",
        "parameters": [
          {
            "description": "修改密码信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdatePasswordReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/user/v1/update/phone": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "修改手机号",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "修改手机号",
        "parameters": [
          {
            "description": "修改手机号信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdatePhoneReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/user/v1/update/profile": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "修改：昵称、头像、签名、背景图片、邮箱",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户模块"
        ],
        "summary": "修改个人信息",
        "parameters": [
          {
            "description": "修改成功",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdateUserProfileReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "修改成功",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    },
    "/gateway/video/feed": {
      "get": {
        "description": "获取推荐视频流",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "视频模块"
        ],
        "summary": "获取视频流",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "最新时间戳（分页）",
            "name": "latest_time",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.VideoInfoRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/video/list": {
      "get": {
        "description": "获取指定用户的视频列表",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "视频模块"
        ],
        "summary": "获取用户视频列表",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "用户ID",
            "name": "user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 30,
            "description": "每页数量",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.VideoInfoRes"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/gateway/video/publish": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "用户发布视频",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "视频模块"
        ],
        "summary": "发布视频",
        "parameters": [
          {
            "description": "视频信息",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.PublishVideoReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "发布成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/res.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.PublishVideoResp"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "参数错误",
            "schema": {
              "$ref": "#/definitions/res.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "dto.LoginReq": {
      "type": "object",
      "required": [
        "password",
        "username"
      ],
      "properties": {
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "123456"
        },
        "username": {
          "type": "string",
          "example": "admin"
        }
      }
    },
...
    "dto.PublishVideoReq": {
      "type": "object",
      "required": [
        "cover_url",
        "play_url",
        "title"
      ],
      "properties": {
        "cover_url": {
          "type": "string"
        },
        "description": {
          "type": "string",
          "maxLength": 1000
        },
        "duration": {
          "type": "number",
          "minimum": 0
        },
        "height": {
          "type": "integer",
          "minimum": 0
        },
        "play_url": {
          "type": "string"
        },
        "title": {
          "type": "string",
          "maxLength": 255
        },
        "width": {
          "type": "integer",
          "minimum": 0
        }
      }
    },
    "dto.PublishVideoResp": {
      "type": "object",
      "properties": {
        "video_id": {
          "type": "string",
          "example": "0"
        }
      }
    },
    "dto.RegisterReq": {
      "type": "object",
      "required": [
        "password",
        "phone",
        "username"
      ],
      "properties": {
        "email": {
          "description": "Email 注册时可选，但如果传了必须符合格式",
          "type": "string",
          "example": "test@example.com"
        },
        "password": {
          "type": "string",
          "maxLength": 32,
          "minLength": 6,
          "example": "secret123"
        },
        "phone": {
          "type": "string",
          "example": "18888888888"
        },
        "username": {
          "type": "string",
          "maxLength": 20,
          "minLength": 4,
          "example": "newuser"
        }
      }
    },
...
    "dto.UpdatePasswordReq": {
      "type": "object",
      "required": [
        "newPassword",
        "oldPassword"
      ],
      "properties": {
        "newPassword": {
          "description": "nefield 确保新旧密码不等",
          "type": "string",
          "maxLength": 32,
          "minLength": 6
        },
        "oldPassword": {
          "type": "string"
        }
      }
    },
    "dto.UpdatePhoneReq": {
      "type": "object",
      "required": [
        "code",
        "newPhone"
      ],
      "properties": {
        "code": {
          "description": "增加验证码字段，实际业务必需",
          "type": "string"
        },
        "newPhone": {
          "type": "string"
        }
      }
    },
    "dto.UpdateUserProfileReq": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "backgroundImage": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "nickName": {
          "type": "string",
          "maxLength": 50
        },
        "signature": {
          "type": "string",
          "maxLength": 255
        }
      }
    },
    "dto.UserInfoRes": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string"
        },
        "backgroundImage": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "nickName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "signature": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
...
    "dto.VideoInfoRes": {
      "type": "object",
      "properties": {
        "author": {
          "description": "作者信息",
          "allOf": [
            {
              "$ref": "#/definitions/dto.UserInfoRes"
            }
          ]
        },
        "comment_count": {
          "type": "integer"
        },
        "cover_url": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "number"
        },
        "favorite_count": {
          "type": "integer"
        },
        "height": {
          "type": "integer"
        },
        "id": {
          "type": "string",
          "example": "0"
        },
        "is_favorite": {
          "description": "当前用户是否点赞",
          "type": "boolean"
        },
        "play_url": {
          "type": "string"
        },
        "publish_time": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "user_id": {
          "type": "string",
          "example": "0"
        },
        "view_count": {
          "type": "integer"
        },
        "width": {
          "type": "integer"
        }
      }
    },
    "res.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    }
  }
}